Add the "Step" spin box
The default behavior of LPCG is to generate notes that test you on all lines of the poem
(All lines eventually appear in the Line field).

This option allows customizing that by allowing to change the "gap" (or step)
between the first line of the Line field of each successive note.

A step of 1 will result in the usual behavior.